Make repo compatible with Please v17
Please 17.0.0-beta.6 removed the built-in Go rules, so the version
constraint in `.plzconfig` is no longer valid - require Please 17.0.0 at
a minimum, and explicitly load the Go rules from the `go-rules` repo.

This also requires the in-repo Go toolchain to be bumped to 1.19, since
the version of `please_go` in `go-rules` 1.6.0 uses features added in
the Go 1.19 standard library.