Explicitly make `zip` preamble options mutually exclusive
The `--preamble`, `--preamble_from` and `--preamble_file` options to the
`zip` command exhibit strange behaviour when more than one of them is
specified. If `--preamble_file` and either of the other two options are
given, the preamble from `--preamble`/`--preamble_from` will be written
first, followed by the contents of `--preamble_file`. If `--preamble`
and `--preamble_from` are given, only `--preamble_from` is honoured. If
all three are specified, only `--preamble` and `--preamble_file` are
honoured.

It really only makes sense to use at most one of these options during a
single invocation of the `zip` command, so exit in failure if more than
one is specified.