`zip`: read all non-zip data from `--preamble_from` file
The `zip` command's `--preamble_from` option, as used by Please's
python-rules plugin, has strange semantics. The help text says it reads
the first line from the given file and uses it as the preamble for the
generated zip file, which is true unless the first line is `#!/bin/sh`,
in which case the second line will also be read. The second line of
"main" pex files generated by python-rules contain an `exec` command,
allowing them to self-execute with more arguments than would otherwise
be supported in a traditional shebang.

Genericise the behaviour of `--preamble_from` so that it is useful for
more than just files that begin with 1 or 2-line shell commands - assume
the input is a zip file (a tacit assumption until now), and read all
non-zip data from the start of the file and use it as the preamble for
the generated zip file.

Because python-rules' pex files are either pure zip files or zip files
prepended with a 1 or 2-line shell command, this does not break
backwards compatibility with python-rules.